Implemented @color command
@color command created to allow players to manage their color settings
and choose the best color options available for their client.

Altered the test patterns in $m/lib/color so that the color test itself
is at the bottom rather than the top, so it's not lost in the scrollback
if somebody is too lazy (or concevably if their screen is too small and
has no convenient scrollback) to scroll up.

Added M-LIB-COLOR-encoding_player_valid to list allowed player output
encodings.