Created a manual color matching algorithm for 4BIT ANSI
I wasn't satisfied with the quality I was getting out of the color space
matching. It felt a little TOO accurate. The differences in paleness
between the VGA and the XTerm palette were causing a lot of pale colors
to go in strange directions toward whites and grays or not depending on
which palette you were using.

But the colors are on the extremes of the color wheel for a reason. You
can neatly chop up the color space into a comfy little pie for them so
that's what I did and it looks a lot better.