Completed more of the 'sense' program
Added properties to let you use it to see exits and contents, and
removed the 'also sense everyhing in the room' property. Now it should
be suitable to use as a 'look' program as well as the other senses.

It now uses the object substitution routine in $m/lib/grammar.

Also that routine now supports using more than one object at once in the
substitution.

Also created an oxford comma list join routine in the process.