Continued implementation of the senses command
Renamed the properties.

Made it use the new verb functionality in $m/lib/grammar substitution.

Made it use the new object theme coloring functionality as well.

Removed target notification specifically. It just notifies the room in
general, including the player that used the action.

Made it respect DARK on the room and the objects, and SILENT on the
player.

Made it use the error lines from $m/lib/theme