Introduced custom string type callbacks for $m/lib/string
Now $m/lib/string routines can handle any kind of color-coded string, or
even strings with other kids of metadata in them, by allowing the
caller to define a few callbacks to handle basic operations like
splitting and concatenating.

Callback versions of loads of built-in string primitives have been
created as well.

Also created a hard/soft distinction for M-LIB-COLOR-strcat, that allows
or prevents colors to bleed from the end of the first string to the
beginning of the next. Effectively it's just sticking a reset code in
the middle, but in principle it might do more if more complex color
codes are added?